Refactor test folders for ed448 (#364)
